Changing argument of constructor vom int to long since int eventually always ends up being too small.
git-svn-id: file:///srv/local/data/cleanRepo/trunk@3695 a0517364-e873-4167-b1c8-416daa5a6b91

+++ matsim/src/org/matsim/basic/v01/IdImpl.java
@@ -33,8 +33,8 @@ public class IdImpl implements Id, Serializable {
 		this.id = id;
 	}
 
-	public IdImpl(final int id) {
-		this.id = Integer.toString(id);
+	public IdImpl(final long id) {
+		this.id = Long.toString(id);
 	}
 
 	@Override
